Reorganized components into basic, partials, and sections folders
