+++ src/components/partials/footer/Footer.tsx
@@ -1,12 +1,12 @@
 import Link from 'next/link';
 
-import { Background } from '../background/Background';
-import { FooterCompanyInfo } from '../footer/FooterCompanyInfo';
-import { FooterLinks } from '../footer/FooterLinks';
-import { FooterTwoRowsCopyright } from '../footer/FooterTwoRowsCopyright';
-import { Section } from '../layout/Section';
-import { AppConfig } from '../../utils/AppConfig';
-import { Logo } from './Logo';
+import { Background } from '../../basic/background/Background';
+import { FooterCompanyInfo } from './FooterCompanyInfo';
+import { FooterLinks } from './FooterLinks';
+import { FooterTwoRowsCopyright } from './FooterTwoRowsCopyright';
+import { Section } from '../../basic/section/Section';
+import { AppConfig } from '../../../utils/AppConfig';
+import { Logo } from '../logo/Logo';
 
 const Footer = () => (
   <Background color="bg-gray-100">
