+++ src/components/partials/footer/FooterTwoRowsCopyright.tsx
