+++ src/components/partials/logo/Logo.tsx
@@ -1,4 +1,4 @@
-import { AppConfig } from '../../utils/AppConfig';
+import { AppConfig } from '../../../utils/AppConfig';
 
 type ILogoProps = {
   xl?: boolean;
