+++ src/components/partials/meta/Meta.tsx
@@ -1,7 +1,7 @@
 import { NextSeo } from 'next-seo';
 import Head from 'next/head';
 
-import { AppConfig } from '../../utils/AppConfig';
+import { AppConfig } from '../../../utils/AppConfig';
 
 type IMetaProps = {
   title: string;
